+++ services/web/server/src/simcore_service_webserver/folders/_models.py
@@ -1,0 +1,69 @@
+import logging
+
+from models_library.basic_types import IDStr
+from models_library.folders import FolderID
+from models_library.rest_ordering import OrderBy, OrderDirection
+from models_library.rest_pagination import PageQueryParameters
+from models_library.users import UserID
+from models_library.utils.common_validators import null_or_none_str_to_none_validator
+from models_library.workspaces import WorkspaceID
+from pydantic import Extra, Field, Json, validator
+from servicelib.aiohttp.requests_validation import RequestParams, StrictRequestParams
+from servicelib.request_keys import RQT_USERID_KEY
+
+from .._constants import RQ_PRODUCT_KEY
+
+_logger = logging.getLogger(__name__)
+
+
+class FoldersRequestContext(RequestParams):
+    user_id: UserID = Field(..., alias=RQT_USERID_KEY)  # type: ignore[literal-required]
+    product_name: str = Field(..., alias=RQ_PRODUCT_KEY)  # type: ignore[literal-required]
+
+
+class FoldersPathParams(StrictRequestParams):
+    folder_id: FolderID
+
+
+class FolderListWithJsonStrQueryParams(PageQueryParameters):
+    # pylint: disable=unsubscriptable-object
+    order_by: Json[OrderBy] = Field(
+        default=OrderBy(field=IDStr("modified"), direction=OrderDirection.DESC),
+        description="Order by field (modified_at|name|description) and direction (asc|desc). The default sorting order is ascending.",
+        example='{"field": "name", "direction": "desc"}',
+        alias="order_by",
+    )
+    folder_id: FolderID | None = Field(
+        default=None,
+        description="List the subfolders of this folder. By default, list the subfolders of the root directory (Folder ID is None).",
+    )
+    workspace_id: WorkspaceID | None = Field(
+        default=None,
+        description="List folders in specific workspace. By default, list in the user private workspace",
+    )
+
+    @validator("order_by", check_fields=False)
+    @classmethod
+    def validate_order_by_field(cls, v):
+        if v.field not in {
+            "modified_at",
+            "name",
+            "description",
+        }:
+            msg = f"We do not support ordering by provided field {v.field}"
+            raise ValueError(msg)
+        if v.field == "modified_at":
+            v.field = "modified"
+        return v
+
+    class Config:
+        extra = Extra.forbid
+
+    # validators
+    _null_or_none_str_to_none_validator = validator(
+        "folder_id", allow_reuse=True, pre=True
+    )(null_or_none_str_to_none_validator)
+
+    _null_or_none_str_to_none_validator2 = validator(
+        "workspace_id", allow_reuse=True, pre=True
+    )(null_or_none_str_to_none_validator)
